pyproject.toml
[project]
name = "mediaskinoscope"
version = "0.1.0"
description = "Real-time dematology alerts over live-streams."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "gradio>=5.38.2",
    "llvmlite>=0.42.0",
    "numba>=0.59.0",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "fastrtc>=0.0.29",
    "sentencepiece>=0.2.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers[torch]>=4.53.2",
    "pipecat-ai>=0.0.76",
]
